fix(api): use MikroORM folder discovery
