+++ apps/api/src/modules/auth/entities/account.entity.ts
@@ -1,0 +1,49 @@
+import {
+  Entity,
+  ManyToOne,
+  PrimaryKey,
+  Property,
+} from '@mikro-orm/decorators/legacy'
+import { User } from './user.entity'
+
+@Entity({ tableName: 'account' })
+export class Account {
+  @PrimaryKey({ type: 'uuid', defaultRaw: 'gen_random_uuid()' })
+  id!: string
+
+  @Property()
+  accountId!: string
+
+  @Property()
+  providerId!: string
+
+  @ManyToOne(() => User, { fieldName: 'userId' })
+  user!: User
+
+  @Property({ nullable: true })
+  accessToken?: string
+
+  @Property({ nullable: true })
+  refreshToken?: string
+
+  @Property({ nullable: true })
+  idToken?: string
+
+  @Property({ nullable: true })
+  accessTokenExpiresAt?: Date
+
+  @Property({ nullable: true })
+  refreshTokenExpiresAt?: Date
+
+  @Property({ nullable: true })
+  scope?: string
+
+  @Property({ nullable: true })
+  password?: string
+
+  @Property()
+  createdAt: Date = new Date()
+
+  @Property({ onUpdate: () => new Date() })
+  updatedAt: Date = new Date()
+}
