+++ apps/api/src/modules/auth/entities/user.entity.ts
@@ -1,0 +1,31 @@
+import {
+  Entity,
+  PrimaryKey,
+  Property,
+  Unique,
+} from '@mikro-orm/decorators/legacy'
+
+@Entity({ tableName: 'user' })
+export class User {
+  @PrimaryKey({ type: 'uuid', defaultRaw: 'gen_random_uuid()' })
+  id!: string
+
+  @Property()
+  name!: string
+
+  @Property()
+  @Unique()
+  email!: string
+
+  @Property()
+  emailVerified: boolean = false
+
+  @Property({ nullable: true })
+  image?: string
+
+  @Property()
+  createdAt: Date = new Date()
+
+  @Property({ onUpdate: () => new Date() })
+  updatedAt: Date = new Date()
+}
